+++ packages/ui/src/index.ts
@@ -43,6 +43,7 @@ export * from "./components/drawer";
 export * from "./components/dropdown-menu";
 export * from "./components/empty-state";
 export * from "./components/field";
+export * from "./components/file-dropzone";
 export * from "./components/grid-list";
 export * from "./components/help-tooltip";
 export * from "./components/hover-card";